Add REST resource for git configuration and connection
