+++ src/main/resources/atlassian-plugin.xml
@@ -120,6 +120,7 @@
 		<!-- ConDec JavaScript files -->
 		<resource type="download" name="general.api.js" location="/js/general.api.js" />
 		<resource type="download" name="condec.api.js" location="/js/condec.api.js" />
+		<resource type="download" name="condec.git.api.js" location="/js/condec.git.api.js" />
 		<resource type="download" name="condec.observable.js" location="/js/condec.observable.js" />
 		<resource type="download" name="condec.dialog.js" location="/js/condec.dialog.js" />
 		<resource type="download" name="condec.context.menu.js" location="/js/condec.context.menu.js" />
