+++ src/main/resources/js/condec.api.js
@@ -472,88 +472,6 @@
 	};
 
 	/*
-	 * external references: settingsForSingleProject.vm,
-	 * settingsForAllProjects.vm
-	 */
-	ConDecAPI.prototype.setKnowledgeExtractedFromGit = function(isKnowledgeExtractedFromGit, projectKey) {
-		generalApi.postJSON(this.restPrefix + "/config/setKnowledgeExtractedFromGit.json?projectKey="
-			+ projectKey + "&isKnowledgeExtractedFromGit=" + isKnowledgeExtractedFromGit, null, function(error,
-				response) {
-			if (error === null) {
-				if (isKnowledgeExtractedFromGit) {
-					showFlag("success", "Git connection for this project is activated.");
-				} else {
-					showFlag("success", "Git connection for this project is deactivated.");
-				}
-			}
-		});
-	};
-
-	/*
-	 * external references: settingsForSingleProject.vm
-	 */
-	ConDecAPI.prototype.setPostFeatureBranchCommits = function(checked, projectKey) {
-		generalApi.postJSON(this.restPrefix + "/config/setPostFeatureBranchCommits.json?projectKey="
-			+ projectKey + "&isPostFeatureBranchCommits=" + checked, null, function(error,
-				response) {
-			if (error === null) {
-				showFlag("success", "Post feature branch commits for this project has been set to " + checked
-					+ ".");
-			}
-		});
-	};
-
-	/*
-	 * external references: settingsForSingleProject.vm
-	 */
-	ConDecAPI.prototype.setPostDefaultBranchCommits = function(checked, projectKey) {
-		generalApi.postJSON(this.restPrefix + "/config/setPostDefaultBranchCommits.json?projectKey="
-			+ projectKey + "&isPostDefaultBranchCommits=" + checked, null, function(error,
-				response) {
-			if (error === null) {
-				showFlag("success", "Post default branch commits for this project has been set to " + checked
-					+ ".");
-			}
-		});
-	};
-
-	/*
-	 * external references: gitSettings.vm
-	 */
-	ConDecAPI.prototype.setGitRepositoryConfigurations = function(projectKey, gitRepositoryConfigurations) {
-		generalApi.postJSON(this.restPrefix + "/config/setGitRepositoryConfigurations.json?projectKey="
-			+ projectKey, gitRepositoryConfigurations, function(error, response) {
-				if (error === null) {
-					showFlag("success", "The git URIs and credentials for this project have been set.");
-				}
-			});
-	};
-
-	/*
-	 * external references: gitSettings.vm
-	 */
-	ConDecAPI.prototype.setCodeFileEndings = function(projectKey, codeFileEndings) {
-		generalApi.postJSON(this.restPrefix + "/config/setCodeFileEndings.json?projectKey="
-			+ projectKey, codeFileEndings, function(error, response) {
-				if (error === null) {
-					showFlag("success", "The code file endings for this project have been set.");
-				}
-			});
-	};
-
-	/*
-	 * external references: gitSettings.vm
-	 */
-	ConDecAPI.prototype.deleteGitRepos = function(projectKey) {
-		generalApi.postJSON(this.restPrefix + "/config/deleteGitRepos.json?projectKey=" + projectKey, null,
-			function(error, response) {
-				if (error === null) {
-					showFlag("success", "The git repos for this project were deleted.");
-				}
-			});
-	};
-
-	/*
 	 * external references: settingsForSingleProject.vm
 	 */
 	ConDecAPI.prototype.setKnowledgeTypeEnabled = function(isKnowledgeTypeEnabled, knowledgeType, projectKey) {
