+++ src/test/java/de/uhd/ifi/se/decision/management/jira/rest/gitrest/TestPostFeatureBranchCommits.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.rest.configrest;
+package de.uhd.ifi.se.decision.management.jira.rest.gitrest;
 
 import static org.junit.Assert.assertEquals;
 
@@ -11,18 +11,18 @@ import org.junit.Test;
 import com.atlassian.jira.mock.servlet.MockHttpServletRequest;
 
 import de.uhd.ifi.se.decision.management.jira.TestSetUp;
-import de.uhd.ifi.se.decision.management.jira.rest.ConfigRest;
+import de.uhd.ifi.se.decision.management.jira.rest.GitRest;
 import de.uhd.ifi.se.decision.management.jira.testdata.JiraUsers;
 
 public class TestPostFeatureBranchCommits extends TestSetUp {
 
 	private HttpServletRequest request;
-	private ConfigRest configRest;
+	private GitRest gitRest;
 
 	@Before
 	public void setUp() {
 		init();
-		configRest = new ConfigRest();
+		gitRest = new GitRest();
 		request = new MockHttpServletRequest();
 		request.setAttribute("user", JiraUsers.SYS_ADMIN.getApplicationUser());
 	}
@@ -30,18 +30,18 @@ public class TestPostFeatureBranchCommits extends TestSetUp {
 	@Test
 	public void testOk() {
 		assertEquals(Status.OK.getStatusCode(),
-				configRest.setPostFeatureBranchCommits(request, "TEST", true).getStatus());
+				gitRest.setPostFeatureBranchCommits(request, "TEST", true).getStatus());
 	}
 
 	@Test
 	public void testInvalidKey() {
 		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				configRest.setPostFeatureBranchCommits(request, null, true).getStatus());
+				gitRest.setPostFeatureBranchCommits(request, null, true).getStatus());
 	}
 
 	@Test
 	public void testEmptyKey() {
 		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				configRest.setPostFeatureBranchCommits(request, "", true).getStatus());
+				gitRest.setPostFeatureBranchCommits(request, "", true).getStatus());
 	}
 }
