+++ src/test/java/de/uhd/ifi/se/decision/management/jira/rest/gitrest/TestSetCodeFileEndings.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.rest.configrest;
+package de.uhd.ifi.se.decision.management.jira.rest.gitrest;
 
 import static org.junit.Assert.assertEquals;
 
@@ -17,47 +17,46 @@ import com.atlassian.jira.mock.servlet.MockHttpServletRequest;
 import de.uhd.ifi.se.decision.management.jira.TestSetUp;
 import de.uhd.ifi.se.decision.management.jira.mocks.MockPluginSettings;
 import de.uhd.ifi.se.decision.management.jira.mocks.MockPluginSettingsFactory;
-import de.uhd.ifi.se.decision.management.jira.rest.ConfigRest;
+import de.uhd.ifi.se.decision.management.jira.rest.GitRest;
 import de.uhd.ifi.se.decision.management.jira.testdata.JiraUsers;
 
 public class TestSetCodeFileEndings extends TestSetUp {
 
 	protected HttpServletRequest request;
-	protected ConfigRest configRest;
-    protected Map<String, String> codeFileEndings;
+	protected GitRest gitRest;
+	protected Map<String, String> codeFileEndings;
 
 	@Before
 	public void setUp() {
 		init();
-		configRest = new ConfigRest();
+		gitRest = new GitRest();
 		request = new MockHttpServletRequest();
 		request.setAttribute("user", JiraUsers.SYS_ADMIN.getApplicationUser());
-        codeFileEndings = new HashMap<String, String>();
-        codeFileEndings.put("JAVA_C", "java");
+		codeFileEndings = new HashMap<String, String>();
+		codeFileEndings.put("JAVA_C", "java");
 	}
 
 	@Test
 	public void testRequestNullProjectKeyNullCodeFileEndingsNull() {
-		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				configRest.setCodeFileEndings(null, null, null).getStatus());
+		assertEquals(Status.BAD_REQUEST.getStatusCode(), gitRest.setCodeFileEndings(null, null, null).getStatus());
 	}
 
 	@Test
 	public void testRequestNullProjectKeyNullCodeFileEndingsProvided() {
 		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				configRest.setCodeFileEndings(null, null, codeFileEndings).getStatus());
+				gitRest.setCodeFileEndings(null, null, codeFileEndings).getStatus());
 	}
 
 	@Test
 	public void testRequestValidProjectKeyValidCodeFileEndingsNull() {
 		assertEquals(Status.BAD_REQUEST.getStatusCode(),
-				configRest.setCodeFileEndings(request, "TEST", null).getStatus());
+				gitRest.setCodeFileEndings(request, "TEST", null).getStatus());
 	}
 
 	@Test
 	public void testEverythingJustFine() {
 		assertEquals(Status.OK.getStatusCode(),
-				configRest.setCodeFileEndings(request, "TEST", codeFileEndings).getStatus());
+				gitRest.setCodeFileEndings(request, "TEST", codeFileEndings).getStatus());
 	}
 
 	@AfterClass
