+++ src/test/java/de/uhd/ifi/se/decision/management/jira/rest/gitrest/TestSetKnowledgeExtractedFromGit.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.rest.configrest;
+package de.uhd.ifi.se.decision.management.jira.rest.gitrest;
 
 import static org.junit.Assert.assertEquals;
 
@@ -17,18 +17,18 @@ import com.atlassian.jira.mock.servlet.MockHttpServletRequest;
 import de.uhd.ifi.se.decision.management.jira.TestSetUp;
 import de.uhd.ifi.se.decision.management.jira.extraction.gitclient.TestSetUpGit;
 import de.uhd.ifi.se.decision.management.jira.extraction.versioncontrol.GitRepositoryConfiguration;
-import de.uhd.ifi.se.decision.management.jira.rest.ConfigRest;
+import de.uhd.ifi.se.decision.management.jira.rest.GitRest;
 import de.uhd.ifi.se.decision.management.jira.testdata.JiraUsers;
 
 public class TestSetKnowledgeExtractedFromGit extends TestSetUp {
 
 	protected HttpServletRequest request;
-	protected ConfigRest configRest;
+	protected GitRest gitRest;
 
 	@Before
 	public void setUp() {
 		init();
-		configRest = new ConfigRest();
+		gitRest = new GitRest();
 		request = new MockHttpServletRequest();
 		request.setAttribute("user", JiraUsers.SYS_ADMIN.getApplicationUser());
 	}
@@ -36,32 +36,32 @@ public class TestSetKnowledgeExtractedFromGit extends TestSetUp {
 	@Test
 	public void testRequestNullProjectKeyNullIsExtractedFalse() {
 		assertEquals(Response.Status.BAD_REQUEST.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(null, null, false).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(null, null, false).getStatus());
 	}
 
 	@Test
 	public void testRequestNullProjectKeyValidIsExtractedFalse() {
 		assertEquals(Response.Status.BAD_REQUEST.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(null, "TEST", false).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(null, "TEST", false).getStatus());
 	}
 
 	@Test
 	public void testRequestValidProjectKeyNullIsExtractedFalse() {
 		assertEquals(Response.Status.BAD_REQUEST.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(request, null, false).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(request, null, false).getStatus());
 	}
 
 	@Test
 	public void testRequestUserNullProjectKeyValidIsExtractedFalse() {
 		request.setAttribute("user", null);
 		assertEquals(Response.Status.UNAUTHORIZED.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(request, "TEST", false).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(request, "TEST", false).getStatus());
 	}
 
 	@Test
 	public void testRequestValidProjectKeyValidIsExtractedFalse() {
 		assertEquals(Response.Status.OK.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(request, "TEST", false).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(request, "TEST", false).getStatus());
 	}
 
 	@Test
@@ -71,9 +71,9 @@ public class TestSetKnowledgeExtractedFromGit extends TestSetUp {
 				"master", "", "", "");
 		gitRepositoryConfigurations.add(gitRepositoryConfiguration);
 		assertEquals(Status.OK.getStatusCode(),
-				configRest.setGitRepositoryConfigurations(request, "TEST", gitRepositoryConfigurations).getStatus());
+				gitRest.setGitRepositoryConfigurations(request, "TEST", gitRepositoryConfigurations).getStatus());
 		assertEquals(Response.Status.OK.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(request, "TEST", true).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(request, "TEST", true).getStatus());
 	}
 
 	@Test
@@ -83,14 +83,14 @@ public class TestSetKnowledgeExtractedFromGit extends TestSetUp {
 		List<GitRepositoryConfiguration> badGitRepositoryConfigurations = new ArrayList<>();
 		badGitRepositoryConfigurations.add(badGitRepositoryConfiguration);
 		assertEquals(Status.OK.getStatusCode(),
-				configRest.setGitRepositoryConfigurations(request, "TEST", badGitRepositoryConfigurations).getStatus());
+				gitRest.setGitRepositoryConfigurations(request, "TEST", badGitRepositoryConfigurations).getStatus());
 		assertEquals(Status.INTERNAL_SERVER_ERROR.getStatusCode(),
-				configRest.setKnowledgeExtractedFromGit(request, "TEST", true).getStatus());
+				gitRest.setKnowledgeExtractedFromGit(request, "TEST", true).getStatus());
 		List<GitRepositoryConfiguration> gitRepositoryConfigurations = new ArrayList<>();
 		GitRepositoryConfiguration gitRepositoryConfiguration = new GitRepositoryConfiguration(TestSetUpGit.GIT_URI,
 				"master", "", "", "");
 		gitRepositoryConfigurations.add(gitRepositoryConfiguration);
 		assertEquals(Status.OK.getStatusCode(),
-				configRest.setGitRepositoryConfigurations(request, "TEST", gitRepositoryConfigurations).getStatus());
+				gitRest.setGitRepositoryConfigurations(request, "TEST", gitRepositoryConfigurations).getStatus());
 	}
 }
